feat: apply new logging system => github.com/devproje/plog
