+++ go.mod
@@ -4,7 +4,7 @@ go 1.19
 
 require (
 	github.com/bwmarrin/discordgo v0.26.1
-	github.com/sirupsen/logrus v1.9.0
+	github.com/devproje/plog v1.0.1
 )
 
 require (
